pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psr-output-analysis-mcp"
version = "0.0.4"
description = "MCP server for Analyse SDDP Outputs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0",
    "pandas"
]

[project.scripts]
mcp-psr-output-analysis = "psr.outputanalysismcp.server:main"

[project.optional-dependencies]
agent = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["psr.outputanalysismcp*", "sddp_agent*"]
